map search

+++ www/wp-content/themes/Zephyr-hodes-ST/src/Map.php
@@ -37,14 +37,37 @@ class Map extends VCComponent {
 
 		// save markers in sql table for nearest search function
 		global $wpdb;
-		$wpdb->query( "TRUNCATE TABLE markers ");
-		
-		$i=0;
-		foreach($locations as $loc){
-			$i++;
-			$sql = $wpdb->prepare("INSERT INTO markers (id, name, address, lat, lng) VALUES ('%d','%s','%s','%s','%s')",$i,$loc['city'],$loc['address'],$loc['latitude'],$loc['longitude']);
-			$wpdb->query($sql);
+		$table_name = 'markers';
+		if($wpdb->get_var("SHOW TABLES LIKE '$table_name'") != $table_name) {
+		     //table not in database. Create new table
+		     $charset_collate = $wpdb->get_charset_collate();
+		 
+				$sql = "CREATE TABLE markers (
+				  id INT NOT NULL AUTO_INCREMENT PRIMARY KEY ,
+				  name VARCHAR( 60 ) NOT NULL ,
+				  address VARCHAR( 80 ) NOT NULL ,
+				  lat FLOAT( 10, 6 ) NOT NULL ,
+				  lng FLOAT( 10, 6 ) NOT NULL,
+				  type VARCHAR( 60 ) NOT NULL 
+				) ENGINE = MYISAM ";
+				$wpdb->query($sql); 
 		}
+		else{
+
+			$wpdb->query( "TRUNCATE TABLE markers ");
+		}
+
+			$i=0;
+			foreach($locations as $loc){
+
+				$i++;
+				if($loc['type']!='placement'){
+					$sql = $wpdb->prepare("INSERT INTO markers (id, name, address, lat, lng, type) VALUES ('%d','%s','%s','%s','%s','%s')",$i,$loc['city'],$loc['address'],$loc['latitude'],$loc['longitude'],$loc['type']);
+					$wpdb->query($sql);
+				}
+			}
+
+
 
 		
 	}
@@ -99,25 +122,23 @@ class Map extends VCComponent {
 				<label><input type="checkbox" checked data-layer="centre" /> <?= __( 'Centres', 'fdm' ) ?></label>
 				<label><input type="checkbox" checked data-layer="academy" /> <?= __( 'Academies', 'fdm' ) ?></label>
 				
-			</div>
-			 <div>
-         <label for="raddressInput">Search location:</label>
-         <input type="text" id="addressInput" size="15"/>
-        <label for="radiusSelect">Radius:</label>
-        <select id="radiusSelect" label="Radius">
-          <option value="50" selected>50 kms</option>
-          <option value="30">30 kms</option>
-          <option value="20">20 kms</option>
-          <option value="10">10 kms</option>
-        </select>
-
+			
+			 <div class="fdm-map-search">
+         
+         <input type="text" id="addressInput" size="15" placeholder="Location"/>
+       
         <input type="button" id="searchButton" value="Search"/>
+        <div id="maptext"></div>
+    </div>
     </div>
 			<div class="fdm-map"></div>
 			
-			<?php foreach( $locations as $loc ) { ?>
+			<?php 
+$i=0;
+			foreach( $locations as $loc ) {  
+				$i++; ?>
 
-				<div class="fdm-map-location" data-marker-src="<?= asset_url( 'img/'.$loc['type'].'-marker.png' ) ?>" data-layer="<?= $loc['type'] ?>" data-latitude="<?= $loc['latitude'] ?>" data-longitude="<?= $loc['longitude'] ?>">
+				<div class="fdm-map-location" data-marker-src="<?= asset_url( 'img/'.$loc['type'].'-marker.png' ) ?>" data-id="<?=$i ?>" data-layer="<?= $loc['type'] ?>" data-latitude="<?= $loc['latitude'] ?>" data-longitude="<?= $loc['longitude'] ?>">
 					<span name="city"><?= $loc['city'] ?></span>
 					<?php if ( $loc['type'] != 'placement' ) { ?>
 						<span name="description">
